Moves client fixture to test_session
